+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 # Changelog
 Changes to this project are documented in this file. More detail and links can be found in the Declarative Onboarding [Document Revision History](https://clouddocs.f5.com/products/extensions/f5-declarative-onboarding/latest/revision-history.html).
 
+## 1.20.0
+### Added
+### Fixed
+- AUTOTOOL-2080: Task status change after restnoded process restarted
+### Changed
+### Removed
+
 ## 1.19.0
 ### Added
 - AUTOTOOL-1990: Route Map